requirements.txt
fastapi
uvicorn
boto3
python-dotenv
python-multipart
requests
httpx
pinecone
sentence-transformers
torch
langchain-text-splitters
pydantic
sarvamai
playwright

